// src/app/api/models/register-user-in.ts
/* tslint:disable */
export interface RegisterUserIn {
  firstName?: string;
  middleName?: string;
  lastName?: string;
  email?: string;
  grade?: string;
  section?: string;
  department?: string;
  phoneNumber?: string;
  username?: string;
  password?: string;
  group?: number;
  authorities?: string;
}